+++ apis/v1alpha1/instrumentation_webhook.go
@@ -31,6 +31,7 @@ const (
 	AnnotationDefaultAutoInstrumentationNodeJS = "instrumentation.opentelemetry.io/default-auto-instrumentation-nodejs-image"
 	AnnotationDefaultAutoInstrumentationPython = "instrumentation.opentelemetry.io/default-auto-instrumentation-python-image"
 	envPrefix                                  = "OTEL_"
+	envSplunkPrefix                            = "SPLUNK_"
 )
 
 // log is for logging in this package.
@@ -80,19 +81,19 @@ var _ webhook.Validator = &Instrumentation{}
 
 // ValidateCreate implements webhook.Validator so a webhook will be registered for the type.
 func (in *Instrumentation) ValidateCreate() error {
-	opentelemetrycollectorlog.Info("validate create", "name", in.Name)
+	instrumentationlog.Info("validate create", "name", in.Name)
 	return in.validate()
 }
 
 // ValidateUpdate implements webhook.Validator so a webhook will be registered for the type.
 func (in *Instrumentation) ValidateUpdate(old runtime.Object) error {
-	opentelemetrycollectorlog.Info("validate update", "name", in.Name)
+	instrumentationlog.Info("validate update", "name", in.Name)
 	return in.validate()
 }
 
 // ValidateDelete implements webhook.Validator so a webhook will be registered for the type.
 func (in *Instrumentation) ValidateDelete() error {
-	opentelemetrycollectorlog.Info("validate delete", "name", in.Name)
+	instrumentationlog.Info("validate delete", "name", in.Name)
 	return nil
 }
 
@@ -130,8 +131,8 @@ func (in *Instrumentation) validate() error {
 
 func (in *Instrumentation) validateEnv(envs []corev1.EnvVar) error {
 	for _, env := range envs {
-		if !strings.HasPrefix(env.Name, envPrefix) {
-			return fmt.Errorf("env name should start with \"OTEL_\": %s", env.Name)
+		if !strings.HasPrefix(env.Name, envPrefix) && !strings.HasPrefix(env.Name, envSplunkPrefix) {
+			return fmt.Errorf("env name should start with \"OTEL_\" or \"SPLUNK_\": %s", env.Name)
 		}
 	}
 	return nil