+++ pkg/collector/upgrade/v0_9_0.go
@@ -19,13 +19,14 @@ import (
 	"strings"
 
 	"gopkg.in/yaml.v2"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/collector/adapters"
+
+	corev1 "k8s.io/api/core/v1"
 )
 
-func upgrade0_9_0(cl client.Client, otelcol *v1alpha1.OpenTelemetryCollector) (*v1alpha1.OpenTelemetryCollector, error) {
+func upgrade0_9_0(u VersionUpgrade, otelcol *v1alpha1.OpenTelemetryCollector) (*v1alpha1.OpenTelemetryCollector, error) {
 	if len(otelcol.Spec.Config) == 0 {
 		return otelcol, nil
 	}
@@ -46,7 +47,9 @@ func upgrade0_9_0(cl client.Client, otelcol *v1alpha1.OpenTelemetryCollector) (*
 			case map[interface{}]interface{}:
 				// delete is a noop if there's no such entry
 				delete(exporter, "reconnection_delay")
-				otelcol.Status.Messages = append(otelcol.Status.Messages, fmt.Sprintf("upgrade to v0.9.0 removed the property reconnection_delay for exporter %q", k))
+				existing := &corev1.ConfigMap{}
+				updated := existing.DeepCopy()
+				u.Recorder.Event(updated, "Normal", "Upgrade", fmt.Sprintf("upgrade to v0.9.0 removed the property reconnection_delay for exporter %q", k))
 				exporters[k] = exporter
 			case string:
 				if len(exporter) == 0 {