+++ pkg/collector/upgrade/versions.go
@@ -16,12 +16,11 @@ package upgrade
 
 import (
 	"github.com/Masterminds/semver/v3"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
 )
 
-type upgradeFunc func(cl client.Client, otelcol *v1alpha1.OpenTelemetryCollector) (*v1alpha1.OpenTelemetryCollector, error)
+type upgradeFunc func(u VersionUpgrade, otelcol *v1alpha1.OpenTelemetryCollector) (*v1alpha1.OpenTelemetryCollector, error)
 
 type otelcolVersion struct {
 	semver.Version
@@ -69,6 +68,10 @@ var (
 		{
 			Version: *semver.MustParse("0.41.0"),
 			upgrade: upgrade0_41_0,
+		},
+		{
+			Version: *semver.MustParse("0.43.0"),
+			upgrade: upgrade0_43_0,
 		},
 	}